+++ gitopscli/commands/sync_apps.py
@@ -3,7 +3,7 @@ import os
 from dataclasses import dataclass
 from typing import Any, Set, Tuple
 from gitopscli.git_api import GitApiConfig, GitRepo, GitRepoApiFactory
-from gitopscli.io.yaml_util import merge_yaml_element, yaml_file_load
+from gitopscli.io_api.yaml_util import merge_yaml_element, yaml_file_load
 from gitopscli.gitops_exception import GitOpsException
 from .command import Command
