+++ tests/commands/test_deploy.py
@@ -7,7 +7,7 @@ import pytest
 from gitopscli.gitops_exception import GitOpsException
 from gitopscli.commands.deploy import DeployCommand
 from gitopscli.git_api import GitRepoApi, GitProvider, GitRepoApiFactory, GitRepo
-from gitopscli.io.yaml_util import update_yaml_file, YAMLException
+from gitopscli.io_api.yaml_util import update_yaml_file, YAMLException
 from .mock_mixin import MockMixin
