+++ tests/commands/test_sync_apps.py
@@ -4,7 +4,7 @@ import unittest
 from unittest.mock import call
 from gitopscli.git_api import GitProvider, GitRepo, GitRepoApi, GitRepoApiFactory
 from gitopscli.commands.sync_apps import SyncAppsCommand
-from gitopscli.io.yaml_util import merge_yaml_element, yaml_file_load
+from gitopscli.io_api.yaml_util import merge_yaml_element, yaml_file_load
 from gitopscli.gitops_exception import GitOpsException
 from .mock_mixin import MockMixin
